backend/requirements.txt
fastapi==0.115.0
PyYAML==6.0.2
uvicorn[standard]==0.30.6
httpx==0.27.2
python-dotenv==1.0.1
pydantic==2.9.2
duckduckgo-search==6.3.7
